requirements.txt
# Autonomous AI Agent Dependencies
# Google Gemini 2.0 API
google-genai

# Screen interaction and automation
pyautogui
mss
pillow

# Windows UI Automation (Accessibility API)
uiautomation

# Web automation (Playwright for precise browser control)
playwright

# Data processing
numpy

# Environment and security
python-dotenv
